Move invoke methods to dedicated file
Co-locate audio-listener api in recording-inputs feature
