+++ src/api/windows.ts
@@ -1,0 +1,25 @@
+import { invoke } from "@tauri-apps/api/core";
+
+import { Commands } from "../types/api";
+
+export const initStandaloneListBox = () => {
+  void invoke(Commands.InitStandaloneListBox);
+};
+
+export const hideStartRecordingDock = () => {
+  void invoke(Commands.HideStartRecordingDock);
+};
+
+type ShowStandaloneListBoxProps = { width: number; x: number; y: number };
+export const showStandaloneListBox = async ({
+  width,
+  x,
+  y,
+}: ShowStandaloneListBoxProps) => {
+  await invoke(Commands.ShowStandaloneListBox, {
+    height: 100,
+    width,
+    x,
+    y,
+  });
+};
