+++ src/main.tsx
@@ -1,13 +1,11 @@
-import { invoke } from "@tauri-apps/api/core";
 import React from "react";
 import { createRoot } from "react-dom/client";
 
+import { initStandaloneListBox } from "./api/windows";
 import { App } from "./app";
-import { Commands } from "./types/api";
-
 import "./index.css";
 
-void invoke(Commands.InitStandaloneListBox);
+initStandaloneListBox();
 
 createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
